Rename project from MB Loader to Origin Client, update package names
