+++ app/src/main/java/com/origin/launcher/Fallback.java
@@ -1,4 +1,4 @@
-package io.bambosan.mbloader;
+package com.origin.launcher;
 
 import android.os.Bundle;
