+++ app/src/main/java/com/origin/launcher/IntentHandler.java
@@ -1,4 +1,4 @@
-package io.bambosan.mbloader;
+package com.origin.launcher;
 
 
 import android.content.Intent;
@@ -19,7 +19,7 @@ public class IntentHandler extends Activity
 		if (isMcLoaded()) {
 			intention.setClassName(getApplicationContext(), "com.mojang.minecraftpe.Launcher");
 		} else {
-			intention.setClassName(getApplicationContext(), "io.bambosan.mbloader.MainActivity");
+			intention.setClassName(getApplicationContext(), "com.origin.launcher.MainActivity");
 		}
 		startActivity(intention);
 	}
