+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">MB Loader</string>
+    <string name="app_name">Origin Client</string>
     <string name="xal_token">ms-xal-0000000048183522</string>
 	<string name="default_string">Hello World!</string>
 	<string name="mbl2_desc">Startup with full support for shader resourcepacks from worlds,realms and servers (might break with mc updates)</string>
