+++ app/src/test/java/com/origin/launcher/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package io.bambosan.mbloader;
+package com.origin.launcher;
 
 import org.junit.Test;
